refactor: Validator를 util/ 하위로 이동
